+++ src/Execution/QueryFilter.php
@@ -8,6 +8,10 @@ use GraphQL\Type\Definition\ResolveInfo;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\ArgFilterDirective;
 
+/**
+ * @deprecated in favour of
+ * @see \Nuwave\Lighthouse\Execution\Builder
+ */
 class QueryFilter
 {
     /**