+++ src/Schema/Directives/FirstDirective.php
@@ -3,7 +3,6 @@
 namespace Nuwave\Lighthouse\Schema\Directives;
 
 use GraphQL\Type\Definition\ResolveInfo;
-use Nuwave\Lighthouse\Execution\QueryFilter;
 use Nuwave\Lighthouse\Schema\Values\FieldValue;
 use Nuwave\Lighthouse\Support\Contracts\FieldResolver;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
@@ -28,19 +27,21 @@ class FirstDirective extends BaseDirective implements FieldResolver
      */
     public function resolveField(FieldValue $fieldValue): FieldValue
     {
+        /** @var \Illuminate\Database\Eloquent\Model $model */
         $model = $this->getModelClass();
 
         return $fieldValue->setResolver(
             function ($root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo) use ($model) {
-                /** @var \Illuminate\Database\Eloquent\Builder $query */
-                $query = QueryFilter::apply(
-                    $model::query(),
-                    $args,
-                    $this->directiveArgValue('scopes', []),
-                    $resolveInfo
-                );
-
-                return $query->first();
+                return $resolveInfo
+                    ->builder
+                    ->addScopes(
+                        $this->directiveArgValue('scopes', [])
+                    )
+                    ->apply(
+                        $model::query(),
+                        $args
+                    )
+                    ->first();
             }
         );
     }