+++ src/Support/Contracts/ArgFilterDirective.php
@@ -2,6 +2,10 @@
 
 namespace Nuwave\Lighthouse\Support\Contracts;
 
+/**
+ * @deprecated in favour of
+ * @see \Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective
+ */
 interface ArgFilterDirective extends ArgDirective
 {
     /**