+++ CHANGELOG.md
@@ -5,11 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/nuwave/lighthouse/compare/v3.2.1...master)
+## [Unreleased](https://github.com/nuwave/lighthouse/compare/v3.4.0...master)
+
+## [3.4.0](https://github.com/nuwave/lighthouse/compare/v3.3.0...v3.4.0) - 2019-04-18
 
 ### Added
 
-- Sync existing models in belongsToMany relations using nested mutations https://github.com/nuwave/lighthouse/pull/707
+- Allow rebinding a custom GlobalId resolver https://github.com/nuwave/lighthouse/pull/739
+
+## [3.3.0](https://github.com/nuwave/lighthouse/compare/v3.2.1...v3.3.0) - 2019-04-15
+
+### Added
+
+- Sync existing models in belongsToMany relations using nested mutations when creating https://github.com/nuwave/lighthouse/pull/707
+- Add `@spread` directive to reshape nested input arguments https://github.com/nuwave/lighthouse/pull/680
+- Add flexible `@builder` directive to quickly specify a single method to apply constraints
+  to the query builder https://github.com/nuwave/lighthouse/pull/680
+- Add `new_between_directives` config to use the new between directives now https://github.com/nuwave/lighthouse/pull/680
+
+### Deprecated
+
+- Use the `@spread` instead of the `flatten` argument of `@create`/`@update` https://github.com/nuwave/lighthouse/pull/680
+- Prefer usage of the `ArgBuilderDirective` instead of the `ArgFilterDirective` https://github.com/nuwave/lighthouse/pull/680
+- `@whereBetween` and `@whereNotBetween` will take a single input object
+  instead of being spread across two args https://github.com/nuwave/lighthouse/pull/680
 
 ### Changed